refactor: common components moved to /components/common folder
