--- src/components/AbsoluteCard/index.ts
@@ -1,1 +1,0 @@
-export { default as AbsoluteCard } from "@/components/AbsoluteCard/AbsoluteCard";
