--- src/components/Card/index.tsx
@@ -1,1 +1,0 @@
-export { default as Card } from "@/components/Card/Card";
