+++ src/components/SignIn/utils/getOAuthIcon.tsx
@@ -1,5 +1,5 @@
 import assertNever from "@/utils/assertNever";
-import { GitHubIcon } from "@/components/Icons";
+import { GitHubIcon } from "@/components/common/Icons";
 import { TAvailableProviders } from "@/components/SignIn";
 
 export function getOAuthIcon(name: Lowercase<TAvailableProviders>) {
