--- src/components/Testimonial/index.tsx
@@ -1,3 +1,0 @@
-export { default as Testimonial } from "@/components/Testimonial/components/Testimonial";
-
-export type { Testimonial as TTestimonial } from "@/components/Testimonial/types";
