+++ src/components/common/AbsoluteCard/AbsoluteCardContent.tsx
