+++ src/components/common/AbsoluteCard/index.ts
@@ -1,0 +1,1 @@
+export { default as AbsoluteCard } from "@/components/common/AbsoluteCard/AbsoluteCard";
