+++ src/components/common/Card/index.tsx
@@ -1,0 +1,1 @@
+export { default as Card } from "@/components/common/Card/Card";
