+++ src/components/common/Icons/LogoIcon.tsx
