+++ src/components/common/Icons/StarIcon.tsx
