+++ src/components/common/Icons/index.tsx
@@ -1,0 +1,14 @@
+export { ManIcon } from "@/components/common/Icons/ManIcon";
+export { HideIcon } from "@/components/common/Icons/HideIcon";
+export { StarIcon } from "@/components/common/Icons/StarIcon";
+export { UserIcon } from "@/components/common/Icons/UserIcon";
+export { CartIcon } from "@/components/common/Icons/CartIcon";
+export { LogoIcon } from "@/components/common/Icons/LogoIcon";
+export { WomanIcon } from "@/components/common/Icons/WomanIcon";
+export { HeartIcon } from "@/components/common/Icons/HeartIcon";
+export { ArrowIcon } from "@/components/common/Icons/ArrowIcon";
+export { UnisexIcon } from "@/components/common/Icons/UnisexIcon";
+export { LogoutIcon } from "@/components/common/Icons/LogoutIcon";
+export { SearchIcon } from "@/components/common/Icons/SearchIcon";
+export { GitHubIcon } from "@/components/common/Icons/GitHubIcon";
+export { ChevronIcon } from "@/components/common/Icons/ChevronIcon";
