+++ src/components/common/Testimonial/index.tsx
@@ -1,0 +1,3 @@
+export { default as Testimonial } from "@/components/common/Testimonial/components/Testimonial";
+
+export type { Testimonial as TTestimonial } from "@/components/common/Testimonial/types";
