+++ src/layouts/Header/components/HeaderActions.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 
 import cn from "@/utils/cn";
-import Button from "@/components/Button";
-import { CartIcon, HeartIcon } from "@/components/Icons";
+import Button from "@/components/common/Button";
+import { CartIcon, HeartIcon } from "@/components/common/Icons";
 import { HeaderAuthBtn } from "@/layouts/Header/components/HeaderAuthBtn";
 
 type HeaderActionsProps = {
